requirements.txt
yaml
feedparser
awesome-slugify
